% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_nn_trees.R
\name{match_nn_trees}
\alias{match_nn_trees}
\title{Match trees based on NN distance}
\usage{
match_nn_trees(mapped_trees, detected_trees, distance_type = "3D")
}
\arguments{
\item{mapped_trees}{data.frame of mapped tree tops with at least the 5 following columns (from col 1 to 5 and in order): plot ID, unique tree ID, X, Y, Z. Other columns can be added after the 5th column.}

\item{detected_trees}{data.frame of detected tree tops with at least the 5 following columns (from col 1 to 5 and in order): plot ID, unique tree ID, X, Y, Z. Other columns can be added after the 5th column.}

\item{distance_type}{Either "2D" or "3D".}
}
\description{
Function to match detected tree tops with stem mapped trees. The 2D or 3D distance between each stem-mapped tree and detected tree is calculated and use to match the trees that are the closest to each others
}
