% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_to_XY.R
\name{polar_to_XY}
\alias{polar_to_XY}
\title{Polar to XY coordianates
Convert polar coordinates to cartesian coordinates. Used for example if trees were mapped form plot center using distance and azimuth}
\usage{
polar_to_XY(
  azimuth,
  distance,
  xcenter,
  ycenter,
  azimuth_offset,
  distance_offset
)
}
\arguments{
\item{azimuth}{- Numeric vector of azimuth from plot center}

\item{distance}{- Numeric vector of distance from plot center}

\item{xcenter}{- Cartesian X coordinate of plot center}

\item{ycenter}{- Cartesian Y coordinate of plot center}

\item{azimuth_offset}{- Optional. If the tree is leaning and measurement was from plot center to the stem, azimuth between stem and tree top.}

\item{distance_offset}{- Optional. If the tree is leaning and measurement was from plot center to the stem, distance between stem and tree top.}
}
\description{
Polar to XY coordianates
Convert polar coordinates to cartesian coordinates. Used for example if trees were mapped form plot center using distance and azimuth
}
